mgs_client.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

class MGSClient:
    def __init__(self):
        self.affiliate_id = os.getenv("MGS_AFFILIATE_ID")
        # Note: MGS doesn't have a public API like DMM, 
        # this is a placeholder or should be replaced with a scraper/unofficial API if available.
        # For now, we return empty list to avoid crashing.
        pass

    def search_works(self, keyword, hits=5):
        """
        MGSの作品を検索する (スタブ)
        """
        print(f"MGS search triggered for: {keyword} (Stub)")
        # If you have a real MGS API implementation, please restore it here.
        return []
