requirements.txt
opencv-python-headless
mediapipe==0.10.14
protobuf<5.0.0
numpy
matplotlib
python-dotenv
requests
japanize-matplotlib
python-amazon-paapi
google-generativeai
feedparser
beautifulsoup4
